Move validators to DTO folder
